requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-crontab==0.7.1
djangorestframework==3.15.1
jwt==1.3.1
mysql-connector-python==8.3.0
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
